Move date range to main header and remove settings from data analysis
